requirements.txt
# Use the FastAPI version required by bittensor
fastapi>=0.110.0,<0.111.0
# Keep uvicorn at the version required by bittensor
uvicorn==0.22.0
redis==5.0.1
python-dotenv==1.0.0
bittensor==9.1.0
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1
celery==5.3.1
# Use pydantic version compatible with the new FastAPI
pydantic>=2.3.0,<3.0.0
pydantic-settings>=2.0.0
asyncpg==0.28.0
sqlalchemy==2.0.19
sqlalchemy[asyncio]
bittensor>=5.0.0
aiohttp>=3.8.1
psycopg2-binary==2.9.7

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6